Initial commit
All cool projects need to have a GitHub bot, so let's create our own.

We'll use nodejs for the project since it can be easily deployed to
Heroku, and we'll use Typescript in order to make working with
Javascript more bearable[1].

We are also using the oss-generic Typescript coding guidelines, so let's
setup tslint and editorconfig with the relevant rules to ensure that our
code adheres to those guidelines.

We will be using Heroku to host our bot, so let's add a Procfile as
well, and document setup instructions in the README.adoc.

[1] For more information on what makes Typescript better than plain
    Javascript, search for "Typescript rocks" or "Javascript sucks" on
    the internet.