server: add source-map-support
The typescript compiler generates source maps for us, however node does
not recognize them. As a result, we get error stack traces pointing to
the generated .js files rather than the source .ts files:

    /se-edu-bot/dist/server.js:3
        throw new Error('Some error');
        ^

    Error: Some error
        at Object.<anonymous> (/se-edu-bot/dist/server.js:3:11)
        ...

This makes finding the exact line where the error occured in the source
.ts files harder than it should be.

As such, let's use the source-map-support package, which teaches node to
use source maps to replace the paths and line numbers of source-mapped
files with their original paths and line numbers. Using this package, we
get error stack traces like this instead:

    /se-edu-bot/server.ts:5
        throw new Error('Some error');
              ^
    Error: Some error
        at Object.<anonymous> (/se-edu-bot/server.ts:5:11)
        ...