server: implement support for ".env" files
We are going to be using environment variables to configure the bot, to
store settings which do not belong in version control. Environment
variables are used for configuration because they are strongly supported
by Heroku.

However, working with environment variables during development can be
kind of unwieldy, as we will either need to set them up one by one on
the command line every time we wish to develop se-edu-bot, or modify our
system settings.

It would be nice if we had a file where we could just dump all our
environment variables inside, and server.ts would automatically pick
them up when it runs.

Fortunately, there is a popular package in the nodejs world called
"dotenv" which does just that. It will load environment variables from
the ".env" file in the project root when called to do so. Let's use
that.