server: implement "Hello world!" web server
We'll need to implement a web server to respond to GitHub webhook
events.

As a step towards that goal, let's implement a simple web server using
Koa[1] that responds with "Hello world!".

Koa is used due to the simplicity of its core concepts (every middleware
is a function that returns a promise).

In later commits, we will be building on top of this web server code to
fully flesh out se-edu-bot's behavior, including responding to GitHub
webhook events.

[1] https://github.com/koajs/koa