Implement testing support with mocha
Tests are always good, as they can help weed out bugs and prevent
regressions.

While we don't specifically require our code to be well tested, it is
still good to have a testing framework available at hand so that
developers can implement tests when they want to.

Let's use mocha[1] as the testing framework, and mocha-typescript[2] to
make it pleasant to use in typescript. Set mocha up such that all
*.spec.ts files are automatically run as tests. As an example of how to
implement tests, write tests for server.ts as well.

[1] https://mochajs.org/
[2] https://github.com/pana-cc/mocha-typescript