logic: implement github API support
While logic components can receive GitHub webhook events, they can't
really act on them as they do not have access to the GitHub API.

Give logic components access to the GitHub API by passing them a
RequestApi object which is authorized with the installation of the
GitHub App on the se-edu organization. This is implemented by
implementing two Koa middlewares, koaGhAppApi() and
koaGhInstallationApi(). The former installs a RequestApi on the koa
context which is authorized with the GitHub App, while the latter
installs a RequestApi on the koa context which is authorized with the
installation access token.