lib/github: implement Github webhook support
In order for se-edu-bot to be useful it will need to be aware of events
which occur on GitHub. (e.g. it will need to be aware of when PRs are
merged so that it can post comments to thank authors for their hard
work).

GitHub uses webhooks[1] to notify se-edu-bot of events that we are
interested in. To receive these events, let's implement a webhook
endpoint at /webhook.

As an example of how to implement support for a webhook event type,
let's also implement support for the "ping"[2] event.

[1] https://developer.github.com/webhooks/
[2] https://developer.github.com/webhooks/#ping-event