lib/Auth: implement support for authenticated user sessions
Some functionality of se-edu-bot that we wish to implement in the future
requires se-edu-bot to be aware of the user's identity.

For GitHub Apps which is what se-edu-bot is using, this requires us to
identify users via OAuth[1], which in turn requires us to implement
OAuth endpoints. We will also need to use cookies to keep track of the
user's session so we can persistently identify the user.

[1] https://developer.github.com/apps/building-integrations/setting-up-and-registering-github-apps/identifying-users-for-github-apps/