logs: only allow organization members to access them
The bot's logs are accessible to everyone at the /logs endpoint. This is
dangerous because the logs may contain sensitive info.

Fix this by only allowing se-edu organization members to view the logs.

Note that our access control doesn't actually check if the current user
is a member of se-edu, but rather whether the user has access to the
se-edu-bot installation in the se-edu organization. This should be
equivalent.